agent/pyproject.toml
[project]
name = "pulpoo-voice-agent"
version = "0.1.0"
description = "Voice AI agent for creating tasks in Pulpoo using LiveKit"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents>=1.2.0",
    "livekit-plugins-openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "livekit-api>=0.6.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
